% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth2_objects.R
\name{Jwk}
\alias{Jwk}
\title{Google OAuth2 API Objects 
Obtains end-user authorization grants for use with other Google APIs.}
\usage{
Jwk(Jwk.keys = NULL, keys = NULL)
}
\arguments{
\item{Jwk.keys}{The \link{Jwk.keys} object or list of objects}

\item{keys}{No description}
}
\value{
Jwk object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:24:35
filename: /Users/mark/dev/R/autoGoogleAPI/googleoauth2v2.auto/R/oauth2_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Jwk Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Jwk functions: \code{\link{Jwk.keys}}
}

