% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth2_functions.R
\docType{package}
\name{oauth2_googleAuthR}
\alias{oauth2_googleAuthR}
\alias{oauth2_googleAuthR-package}
\title{Google OAuth2 API
Obtains end-user authorization grants for use with other Google APIs.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:24:35
filename: /Users/mark/dev/R/autoGoogleAPI/googleoauth2v2.auto/R/oauth2_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/plus.login
\item https://www.googleapis.com/auth/plus.me
\item https://www.googleapis.com/auth/userinfo.email
\item https://www.googleapis.com/auth/userinfo.profile
}
}

