% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth2_objects.R
\name{Userinfoplus}
\alias{Userinfoplus}
\title{Userinfoplus Object}
\usage{
Userinfoplus(email = NULL, family_name = NULL, gender = NULL,
  given_name = NULL, hd = NULL, id = NULL, link = NULL, locale = NULL,
  name = NULL, picture = NULL, verified_email = NULL)
}
\arguments{
\item{email}{The user's email address}

\item{family_name}{The user's last name}

\item{gender}{The user's gender}

\item{given_name}{The user's first name}

\item{hd}{The hosted domain e}

\item{id}{The obfuscated ID of the user}

\item{link}{URL of the profile page}

\item{locale}{The user's preferred locale}

\item{name}{The user's full name}

\item{picture}{URL of the user's picture image}

\item{verified_email}{Boolean flag which is true if the email address is verified}
}
\value{
Userinfoplus object
}
\description{
Userinfoplus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
